% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_visualization.R
\name{replace_special_chars}
\alias{replace_special_chars}
\title{Replace special characters and remove non-ASCII characters}
\usage{
replace_special_chars(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with special characters replaced and non-ASCII characters removed
}
\description{
Replace special characters and remove non-ASCII characters
}
