% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_goodreads.R
\name{search_goodreads}
\alias{search_goodreads}
\title{Search Goodreads}
\usage{
search_goodreads(
  search_term,
  search_in = c("title", "author"),
  num_books = 10,
  sort_by = "ratings"
)
}
\arguments{
\item{search_term}{A search term string.}

\item{search_in}{Where to search (e.g., "title", "author").}

\item{num_books}{Number of books to return.}

\item{sort_by}{How to sort the results (e.g., "ratings", "published_year").}
}
\value{
A data frame of search results.
}
\description{
This function searches books on Goodreads.
}
\examples{
search_goodreads("parenting", search_in = "title", num_books = 2)
}
